smart-book-search/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
jinja2==3.1.6
sentence-transformers==5.1.0
faiss-cpu==1.12.0
numpy==2.3.2
openai==1.101.0
python-dotenv==1.1.1